refactor pytest fixtures
