+++ dor/service_layer/workframe.py
@@ -1,0 +1,5 @@
+def workframe(func):
+    def wrap():
+        func()
+
+    return wrap
