+++ features/pytest/fixtures/path_data.py
@@ -1,0 +1,16 @@
+import pytest
+
+from pathlib import Path
+
+from dor.domain.models import PathData
+
+@pytest.fixture
+def path_data() -> PathData:
+    scratch = Path("./features/scratch")
+
+    return PathData(
+        scratch=scratch,
+        inbox=Path("./features/fixtures/inbox"),
+        workspaces=scratch / "workspaces",
+        storage=scratch / "storage"
+    )
